requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7
razorpay==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
